backend/evaluations-database/requirements.txt
grpc-interceptor ~= 0.11.0
grpcio
grpcio-tools ~= 1.30